# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateBody": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "path": {
            "default": "",
            "maxLength": 4096,
            "title": "Path",
            "type": "string"
          },
          "placement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Placement"
              },
              {
                "type": "null"
              }
            ]
          },
          "root": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root"
          }
        },
        "title": "CreateBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "OpenBody": {
        "additionalProperties": false,
        "properties": {
          "read_only": {
            "default": false,
            "title": "Read Only",
            "type": "boolean"
          },
          "take_over": {
            "default": false,
            "title": "Take Over",
            "type": "boolean"
          }
        },
        "title": "OpenBody",
        "type": "object"
      },
...
      "Placement": {
        "enum": [
          "in_folder",
          "split",
          "external"
        ],
        "title": "Placement",
        "type": "string"
      },
...
      "RelinkBody": {
        "additionalProperties": false,
        "description": "Desktop: ``choose_folder`` (absolute). Server: ``root`` and ``path`` inside it, the\nway the folder browser names folders (ADR 0035).",
        "properties": {
          "choose_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choose Folder"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "root": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root"
          }
        },
        "title": "RelinkBody",
        "type": "object"
      },
...
      "RenameBody": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RenameBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MosAic",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/projects": {
      "get": {
        "operationId": "list_projects_api_projects_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Projects Api Projects Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Projects"
      },
      "post": {
        "description": "Create the project (or open it when the folder already is one) and start a scan.",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Project Api Projects Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Project"
      }
    },
...
    "/api/projects/{pid}": {
      "patch": {
        "description": "S0 header and S21 General: the project's name (the folder keeps its own).",
        "operationId": "rename_api_projects__pid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Rename Api Projects  Pid  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename"
      }
    },
...
    "/api/projects/{pid}/benchmark": {
      "post": {
        "description": "Measure this computer (ADR 0030); the job runs in ``pid``'s queue.",
        "operationId": "post_benchmark_api_projects__pid__benchmark_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Post Benchmark Api Projects  Pid  Benchmark Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Benchmark"
      }
    },
...
    "/api/projects/{pid}/close": {
      "post": {
        "operationId": "close_api_projects__pid__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Close Api Projects  Pid  Close Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Close"
      }
    },
...
    "/api/projects/{pid}/open": {
      "post": {
        "operationId": "open__api_projects__pid__open_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/OpenBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Open  Api Projects  Pid  Open Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open "
      }
    },
...
    "/api/projects/{pid}/relink": {
      "post": {
        "description": "Re-scan the project's folder, or the folder it was moved to (``choose_folder``):\nfiles whose content fingerprint matches keep all their analysis (M1 acceptance 3).",
        "operationId": "relink_api_projects__pid__relink_post",
        "parameters": [
          {
            "in": "path",
            "name": "pid",
            "required": true,
            "schema": {
              "title": "Pid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RelinkBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Relink Api Projects  Pid  Relink Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Relink"
      }
    },
...
  }
}